# fix: use a simple formatter for converting PG timestamps