Replace using DateTime with a simple formatter that extracts and normalizes the information from a Postgres timestamptz. The former method significantly impacted performance when formatting thousands of Postgres timestamps.