@lseppala 2018-02-28
Before
```
Running 1m test @ https://staging.conch.joyent.us/workspace/496f76b4-8245-4d41-8d97-42fe988401c5/device
  2 threads and 2 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.91s   413.18ms   4.68s    75.86%
    Req/Sec     0.00      0.00     0.00    100.00%
  Latency Distribution
     50%    3.85s
     75%    4.20s
     90%    4.54s
     99%    4.68s
  29 requests in 1.00m, 27.99MB read
Requests/sec:      0.48
Transfer/sec:    477.27KB
```
After
```
Running 1m test @ https://staging.conch.joyent.us/workspace/496f76b4-8245-4d41-8d97-42fe988401c5/device
  2 threads and 2 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.70s   437.21ms   2.32s    62.32%
    Req/Sec     0.09      0.28     1.00     91.30%
  Latency Distribution
     50%    1.64s
     75%    2.17s
     90%    2.28s
     99%    2.32s
  69 requests in 1.00m, 65.90MB read
Requests/sec:      1.15
Transfer/sec:      1.10MB
```